添加了实时聊天功能

+++ UserService/src/main/java/com/yachiyo/UserService/service/Impl/UserInteractServiceImpl.java
@@ -33,11 +33,15 @@ public class UserInteractServiceImpl implements UserInteractService {
     @Override
     public Mono<Result<Boolean>> follow(Long userId, Long followeeId) {
 
+        if (followeeId.equals(userId)) {
+            return Mono.just(Result.error("400", "不能关注自己", "不能关注自己"));
+        }
+
         FollowLink followLink = new FollowLink();
         followLink.setFollowerId(userId);
         followLink.setFolloweeId(followeeId);
 
-        return followLinkRepository.existsFollowLinkByFolloweeIdAndFollowerId(userId, followeeId)
+        return followLinkRepository.existsByFolloweeIdAndFollowerId(userId, followeeId)
                 .flatMap(exists -> exists ? followLinkRepository.deleteByFolloweeIdAndFollowerId(userId, followeeId) : followLinkRepository.insert(followLink))
                 .map(_ -> Result.success(true))
                 .onErrorResume(error -> Mono.just(Result.error("500", error.getMessage(), "关注失败")));
@@ -74,7 +78,8 @@ public class UserInteractServiceImpl implements UserInteractService {
 
     @Override
     public Mono<Result<Boolean>> getFollowStatus(Long userId, Long followeeId) {
-        return followLinkRepository.existsFollowLinkByFolloweeIdAndFollowerId(userId, followeeId)
+        return followLinkRepository.existsByFolloweeIdAndFollowerId(userId, followeeId)
+                .defaultIfEmpty(false)
                 .map(Result::success)
                 .onErrorResume(error -> Mono.just(Result.error("500", error.getMessage(), "获取关注状态失败")));
     }
@@ -88,9 +93,9 @@ public class UserInteractServiceImpl implements UserInteractService {
         return userDetailRepository.findByUserNameContains(userName, PageRequest.of(pageNum - 1, pageSize))
                 .flatMap(userDetail -> {
                     // 并发查询
-                    Mono<Boolean> isFollowing = followLinkRepository.existsFollowLinkByFolloweeIdAndFollowerId(userDetail.getUserId(), currentUserId);
-                    Mono<Boolean> isFollowed = followLinkRepository.existsFollowLinkByFolloweeIdAndFollowerId(currentUserId, userDetail.getUserId());
-                    Mono<Long> followerCount = followLinkRepository.countFollowLinkByFolloweeId(userDetail.getUserId());
+                    Mono<Boolean> isFollowing = followLinkRepository.existsByFolloweeIdAndFollowerId(userDetail.getUserId(), currentUserId).defaultIfEmpty(false);
+                    Mono<Boolean> isFollowed = followLinkRepository.existsByFolloweeIdAndFollowerId(currentUserId, userDetail.getUserId()).defaultIfEmpty(false);
+                    Mono<Long> followerCount = followLinkRepository.countByFolloweeId(userDetail.getUserId());
 
                     // 组合三个 Mono
                     return Mono.zip(isFollowing, isFollowed, followerCount)
@@ -113,8 +118,8 @@ public class UserInteractServiceImpl implements UserInteractService {
 
     @Override
     public Mono<Result<Boolean>> isFriend(Long currentUserId, Long followeeId) {
-        Mono<Boolean> isFollowingMono = followLinkRepository.existsFollowLinkByFolloweeIdAndFollowerId(currentUserId, followeeId);
-        Mono<Boolean> isFollowedMono = followLinkRepository.existsFollowLinkByFolloweeIdAndFollowerId(followeeId, currentUserId);
+        Mono<Boolean> isFollowingMono = followLinkRepository.existsByFolloweeIdAndFollowerId(currentUserId, followeeId).defaultIfEmpty(false);
+        Mono<Boolean> isFollowedMono = followLinkRepository.existsByFolloweeIdAndFollowerId(followeeId, currentUserId).defaultIfEmpty(false);
 
         return Mono.zip(isFollowingMono, isFollowedMono)
                 .map(tuple -> {
@@ -130,7 +135,7 @@ public class UserInteractServiceImpl implements UserInteractService {
         return getFolloweeList(currentUserId)
                 .map(Result::getData)
                 .flatMap(
-                followeeId -> followLinkRepository.existsFollowLinkByFolloweeIdAndFollowerId(currentUserId, followeeId)
+                followeeId -> followLinkRepository.existsByFolloweeIdAndFollowerId(currentUserId, followeeId).defaultIfEmpty(false)
                         .filter(Boolean::booleanValue)
                         .thenReturn(Result.success(followeeId))
                         .onErrorResume(error -> Mono.just(Result.error("500", error.getMessage(), "获取关注状态失败")))
